Move PrimitiveRequestResult to Common.Data.ViewModels for AB#13189.
