+++ Apps/GatewayApi/src/Services/UserEmailService.cs
@@ -26,7 +26,6 @@ namespace HealthGateway.GatewayApi.Services
     using HealthGateway.Common.ErrorHandling;
     using HealthGateway.Common.Models;
     using HealthGateway.Common.Services;
-    using HealthGateway.Database.Constants;
     using HealthGateway.Database.Delegates;
     using HealthGateway.Database.Models;
     using Microsoft.AspNetCore.Http;
