+++ Apps/GatewayApi/test/unit/Services.Test/UserSMSServiceTests.cs
@@ -18,6 +18,7 @@ namespace HealthGateway.GatewayApi.Test.Services
     using System;
     using System.Linq;
     using HealthGateway.Common.Data.Models;
+    using HealthGateway.Common.Data.ViewModels;
     using HealthGateway.Common.Models;
     using HealthGateway.Common.Services;
     using HealthGateway.Database.Constants;
