+++ Apps/WebClient/src/Server/Services/IUserEmailService.cs
@@ -16,7 +16,7 @@
 namespace HealthGateway.WebClient.Services
 {
     using System;
-    using HealthGateway.Common.Models;
+    using HealthGateway.Common.Data.ViewModels;
 
     /// <summary>
     /// The User Email service.
