+++ Apps/WebClient/src/Server/Services/UserSMSService.cs
@@ -21,6 +21,7 @@ namespace HealthGateway.WebClient.Services
     using System.Text.RegularExpressions;
     using HealthGateway.Common.Data.Constants;
     using HealthGateway.Common.Data.Models;
+    using HealthGateway.Common.Data.ViewModels;
     using HealthGateway.Common.Models;
     using HealthGateway.Common.Services;
     using HealthGateway.Database.Delegates;
